youtube_downloader/requirements.txt
pytube3==9.6.4
tqdm==4.62.3
typing-extensions==4.0.1
